code/requirements.txt
attrs==24.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.9
dnspython==2.7.0
filelock==3.16.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-PyMongo==2.3.0
fonttools==4.54.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
numpy==2.1.1
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
PyJWT==2.10.1
pymongo==4.10.1
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
